09_minikube/example/project/product/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking
#
grpcio==1.53.0
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.53.0
    # via -r requirements.in
mysql-connector-python==8.0.29
    # via -r requirements.in
protobuf==4.22.1
    # via
    #   grpcio-tools
    #   mysql-connector-python

# The following packages are considered to be unsafe in a requirements file:
# setuptools
